% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fncols}
\alias{fncols}
\title{Create columns if they don't exist}
\usage{
fncols(data, cname)
}
\arguments{
\item{data}{Dataset you want to change the columns for.}

\item{cname}{Column names to check if exists.}
}
\value{
Returns a dataframe with the extra columns if they don't exist.
}
\description{
This function is taken from [here](https://stackoverflow.com/a/45858044).
}
\examples{
fncols(iris, "species")
}
