% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{longlat_rep}
\alias{longlat_rep}
\title{Replace coordinate colnames}
\usage{
longlat_rep(x)
}
\arguments{
\item{x}{Colnames referring to long/lat infor if pattern matched.}
}
\value{
Vector with replaced strings if pattern found.
}
\description{
Replace coordinate colnames
}
\examples{
# longlat_rep(c ("sample.coordinates.coordinates.0", "sample.coordinates.0"  ))
}
